2008-08-12 13:01 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/make_xmingw.sh   * harbour/make_xmingwce.sh     ! cleanup and some minor fixes     * synced behavior of both scripts in few places
  * harbour/contrib/rddads/ads1.c
  * harbour/contrib/rddads/adsfunc.c
    ! fixed casting of parameters passed to snprintf()

+++ harbour/contrib/rddads/adsfunc.c
@@ -1614,10 +1614,12 @@ HB_FUNC( ADSVERSION )
    switch( hb_parni( 1 ) /* iVersionType */ )
    {
       case 0:
-         snprintf( ucVersion, sizeof( ucVersion ), "%ld.%ld%c", ulMajor, ulMinor, ucLetter );
+         snprintf( ucVersion, sizeof( ucVersion ), "%lu.%lu%c",
+                   ( ULONG ) ulMajor, ( ULONG ) ulMinor, ucLetter );
          break;
       case 3:
-         snprintf( ucVersion, sizeof( ucVersion ), "%s, v%ld.%ld%c", ( char * ) ucDesc, ulMajor, ulMinor, ucLetter );
+         snprintf( ucVersion, sizeof( ucVersion ), "%s, v%lu.%lu%c",
+                   ( char * ) ucDesc, ( ULONG ) ulMajor, ( ULONG ) ulMinor, ucLetter );
          break;
       default:
          ucVersion[ 0 ] = '\0';
